### 2018-11-12 #1045 @arfon
@nproellochs - could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
