### 2018-11-19 #1087 @whedon
[REVIEW]: ReinforcementLearning: A Package to Perform Model-Free Reinforcement Learning in R
**Submitting author:** @nproellochs (<a href="http://orcid.org/0000-0002-1835-7302">Nicolas Proellochs</a>)
**Repository:** <a href="https://github.com/nproellochs/ReinforcementLearning" target ="_blank">https://github.com/nproellochs/ReinforcementLearning</a>
**Version:** v1.0.2
