### 2018-12-01 #1087 @terrytangyuan
@nproellochs Could you clarify the differences between your package and [reinforcelearn](https://cran.r-project.org/web/packages/reinforcelearn/index.html)? Also, have you done any benchmarking or used the package in certain applications yet?
