### 2018-12-01 #1087 @nproellochs
@terrytangyuan The key differences between our package and the reinforcelearn package are as follows:
1) Different from the reinforcelearn package, our package performs batch reinforcement learning (it allows to learn the optimal behavior of an agent based on sample sequences consisting of states, actions and rewards). Batch reinforcement learning drastically reduces the computation time in most settings.
2) Our package is shipped with the built-in capability to sample experience from a function that defines the dynamics of the environment.
3) Our package incorporates a variety of learning algorithms to improve the performance of (batch) reinforcement learning.

We also tested the suitability of the package in multiple application scenarios. For example, our paper "Negation scope detection in sentiment analysis: Decision support for news-driven trading" makes use of the package to learn negations based on reinforcement learning (the paper is available from https://www.sciencedirect.com/science/article/pii/S0167923616300823).
