### 2019-02-07 #1087 @kbenoit
Sorry for the delay guys, here are my comments. All of my suggestions are below but on the [_JOSE_ option above](https://github.com/openjournals/joss-reviews/issues/1087#issuecomment-454359891), I think it has to do below with the purpose of the package and the need it addresses, which I think needs additional clarification before we can answer this definitively. Do I think the package should be rejected from _JOSS_ because it cannot solve large-scale computer vision problems? No, not unless that is what it claims to solve. So for me, a key unanswered question is exactly what need it claims to meet. See more below.

### Checklist issues

* Does the release version given match the GitHub release (v1.0.2)?

Not exactly, as the GitHub repo has not tagged releases. Suggest you use this very useful facility.  https://help.github.com/articles/creating-releases/.  I can verify that the current master branch on the GitHub repo matches the submitted and the CRAN version.

* **Functionality:** Have the functional claims of the software been confirmed?

I find this uncertain, since in the vignette, there is no actual execution of the code, and no interpretation or discussion of the results or the format in which they are delivered.  How I am to understand the output of `policy(model)`, which is a named `character` vector such as
```r
> head(policy(model))
.XXBB..XB XXBB.B.X. .XBB..BXX BXX...B.. ..XB..... XBXBXB...
     "c1"      "c5"      "c5"      "c4"      "c5"      "c9"
```

* **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

None exist in the repo or in the `README.md`, but these would be easy to add.

* **Software paper**

None of the three checklist items are met in the current version of the paper.  See below for my comments on the statement of need and the purpose of the package.

### Core issues relating to package's purpose

It's not entirely clear what the core purpose of the package is, given the performance limitations acknowledged above.  Either this is a practical solution for applying reinforcement learning to problems that people need to solve, in which case there needs to be a discussion o of its performance limitations and some comparison to other possible solutions, in the paper and in my opinion, in the package vignette and/or repository README.  The current statement of purpose (in the paper) suggests that the package exists to provide an R package for model-free learning, implying that its purpose is to provide a practical tool for applying this model to a broad class of problems.  If so, then some discussion of performance is needed.

Or, this is a purely R-based implementation of reinforcement learning that offers a way to test or take apart this class of machine learning model for primarily educational purposes.  If this is the case, then _JOSE_ is more appropriate.

Either way, but especially in the second scenario, more complete documenation is needed, since any user is going to need to:
* understand the required input format for supplying their own data; and
* understand the output format and how to interpret it.

The man pages cite Sutton and Barto (1998).  Why not implement one of the examples from that text using the package, as a vignette?  Especially if this is primarily a learning tool, this would be really useful, but even if designed as a problem-solving tool, it will greatly help users connect the tool to a textbook that more fully explains parameter choices and provides further examples.

Some options are never explained at all, such as `iter`, which defaults to 1.  When would higher values be useful?

### Minor issues

Not every manual entry has examples, and often has very little discussion.  Why not flesh this out, especially if the purpose is educational?

Admittedly, I am a bit obssessed with naming, but I find the function names unintuitive and inconsistent, either non-descriptive (e.g. `policy`) or inconsistent: compounded nouns (e.g. `epsilonGreedyActionSelection`) or verb-nouns (e.g. `lookupLearningRule`).  Not clear either which are data objects and which are functions.

The R code is nicely written and very readable - so if the purpose is educational then it will serve that well, for those looking at source code. Although: please add a space after `if`.

### R CMD CHECK issues

On the a full check, I am seeing:

- `N  checking top-level files
   Non-standard files/directories found at top level:
     ‘JOSS’`
   (add to `.Rbuildignore`)

- `Warning: roxygen2 requires Encoding: UTF-8` (add this to `DESCRIPTION`)
