### 2019-03-12 e5df55e
* Fixed test scripts for R version 3.5.x
