### 2019-04-11 #1087 @kbenoit
Sorry to be slow on this, review by early next week at the latest. Been snowed under with something else but it's done now.
