### 2019-04-21 e6503a4
* Fixed RNG errors * Added test for NA values in rewards column
