+++ tests/testthat.R
@@ -1,4 +1,6 @@
-library(testthat)
+if(!require(testthat)) stop("testthat required for tests")
 library(ReinforcementLearning)
 
+suppressWarnings(RNGversion("3.5.0"))
 test_check("ReinforcementLearning")
+RNGversion(as.character(getRversion()))
