+++ cran-comments.md
@@ -1,3 +1,8 @@
+## Resubmission
+This is a resubmission. In this version I have:
+
+* Inserted suppressWarnings(RNGversion("3.5.0")) before calling set.seed() in the test scripts
+
 ## Test environments
 * Windows 10, R 3.3.3 (local)
 * ubuntu 12.04 (on travis-ci), R 3.3.3
