### 2019-05-01 #1087 @terrytangyuan
@arfon My feedback and opinion still remains unchanged for the reasons I mentioned earlier. I'd recommend @nproellochs submitting this article to other journals, e.g. JOSE mentioned above.
