### 2019-05-01 #1087 @arfon
> @arfon My feedback and opinion still remain unchanged for the reasons I mentioned earlier. I'd recommend @nproellochs submitting this article to other journals, e.g. JOSE mentioned above.

Understood. Thanks @terrytangyuan.
