### 2019-05-20 #1087 @arfon
> @arfon I am planning to incorporate the feedback next weekend. Is there a deadline for the revision?

No, the timing is up to you.

Please make sure though in your revisions you directly address the feedback from @kbenoit. Specifically:

> Or, this is a purely R-based implementation of reinforcement learning that offers a way to test or take apart this class of machine learning model for primarily educational purposes. If this is the case, then JOSE is more appropriate.

In your response here I think you need to make a strong case for this being a JOSS submission rather than a [JOSE](http://jose.theoj.org) one, especially given both reviewers have the same concern here.
