### 2019-05-24 63ef239
* Updated README * Updated DESCRIPTION * Updated .Rbuildignore
