+++ README.md
@@ -1,25 +1,32 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-Reinforcement Learning
-======================
 
-[![Build Status](https://travis-ci.org/nproellochs/ReinforcementLearning.svg?branch=master)](https://travis-ci.org/nproellochs/ReinforcementLearning) [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/ReinforcementLearning)](https://cran.r-project.org/package=ReinforcementLearning)
+# Reinforcement Learning
 
-**ReinforcementLearning** performs model-free reinforcement learning in R. This implementation enables the learning of an optimal policy based on sample sequences consisting of states, actions and rewards. In addition, it supplies multiple predefined reinforcement learning algorithms, such as experience replay.
+[![Build
+Status](https://travis-ci.org/nproellochs/ReinforcementLearning.svg?branch=master)](https://travis-ci.org/nproellochs/ReinforcementLearning)
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/ReinforcementLearning)](https://cran.r-project.org/package=ReinforcementLearning)
 
-Overview
---------
+**ReinforcementLearning** performs model-free reinforcement learning in
+R. This implementation enables the learning of an optimal policy based
+on sample sequences consisting of states, actions and rewards. In
+addition, it supplies multiple predefined reinforcement learning
+algorithms, such as experience replay.
+
+## Overview
 
 The most important functions of **ReinforcementLearning** are:
 
--   Learning an optimal policy from a fixed set of a priori known transition samples
--   Predefined learning rules and action selection modes
--   A highly customizable framework for model-free reinforcement learning tasks
+  - Learning an optimal policy from a fixed set of a priori known
+    transition samples
+  - Predefined learning rules and action selection modes
+  - A highly customizable framework for model-free reinforcement
+    learning tasks
 
-Installation
-------------
+## Installation
 
-You can easily install the latest version of **ReinforcementLearning** with
+You can easily install the latest version of **ReinforcementLearning**
+with
 
 ``` r
 # Recommended option: download and install latest version from CRAN
@@ -30,16 +37,21 @@ install.packages("ReinforcementLearning")
 devtools::install_github("nproellochs/ReinforcementLearning")
 ```
 
-Usage
------
+## Usage
 
-This section shows the basic functionality of how to perform reinforcement learning. First, load the corresponding package **ReinforcementLearning**.
+This section shows the basic functionality of how to perform
+reinforcement learning. First, load the corresponding package
+**ReinforcementLearning**.
 
 ``` r
 library(ReinforcementLearning)
 ```
 
-The following example shows how to learn a reinforcement learning agent using input data in the form of sample sequences consisting of states, actions and rewards. The result of the learning process is a state-action table and an optimal policy that defines the best possible action in each state.
+The following example shows how to learn a reinforcement learning agent
+using input data in the form of sample sequences consisting of states,
+actions and rewards. The result of the learning process is a
+state-action table and an optimal policy that defines the best possible
+action in each state.
 
 ``` r
 # Generate sample experience in the form of state transition tuples
@@ -77,9 +89,24 @@ print(model)
 #> [1] -263
 ```
 
-License
--------
+## Contributing
 
-**ReinforcementLearning** is released under the [MIT License](https://opensource.org/licenses/MIT)
+If you experience any difficulties with the package, or have
+suggestions, or want to contribute directly, you have the following
+options:
 
-Copyright (c) 2018 Nicolas Pröllochs & Stefan Feuerriegel
[…]
