+++ README.Rmd
@@ -75,8 +75,16 @@ model <- ReinforcementLearning(data, s = "State", a = "Action", r = "Reward",
 print(model)
 ```
 
+## Contributing
+
+If you experience any difficulties with the package, or have suggestions, or want to contribute directly, you have the following options:
+
+* Contact the [maintainer](mailto:nicolas.proellochs@wi.jlug.de) by email.
+* Issues and bug reports: [File a GitHub issue](https://github.com/nproellochs/ReinforcementLearning/issues).
+* Fork the source code, modify, and issue a [pull request](https://help.github.com/articles/creating-a-pull-request-from-a-fork/) through the[project GitHub page](https://github.com/nproellochs/ReinforcementLearning).
+
 ## License
 
 **ReinforcementLearning** is released under the [MIT License](https://opensource.org/licenses/MIT)
 
-Copyright (c) 2018 Nicolas Pröllochs & Stefan Feuerriegel
+Copyright (c) 2019 Nicolas Pröllochs & Stefan Feuerriegel
