+++ DESCRIPTION
@@ -11,7 +11,7 @@ Maintainer: Nicolas Proellochs <nicolas.proellochs@is.uni-freiburg.de>
 Description: Performs model-free reinforcement learning in R. This implementation enables the learning
     of an optimal policy based on sample sequences consisting of states, actions and rewards. In 
     addition, it supplies multiple predefined reinforcement learning algorithms, such as experience 
-    replay.
+    replay. Methodological details can be found in Sutton and Barto (1998) <ISBN:0262039249>.
 License: MIT + file LICENSE
 Depends: R (>= 3.2.0)
 Imports:
@@ -23,5 +23,6 @@ Suggests:
     knitr,
     rmarkdown
 LazyData: TRUE
+Encoding: UTF-8
 RoxygenNote: 6.1.1
 VignetteBuilder: knitr
