### 2019-05-24 da5c5fc
* Renamed function experienceReplay() => replayExperience() * Renamed function epsilonGreedyActionSelection() => selectEpsilonGreedyAction() * Renamed function randomActionSelection() => selectRandomAction() * Renamed function policy() => computePolicy() * Fixed typo in README
