### 2019-05-24 19abad4
* Added LICENSE.md
