+++ cran-comments.md
@@ -2,6 +2,7 @@
 This is a resubmission. In this version I have:
 
 * Inserted suppressWarnings(RNGversion("3.5.0")) before calling set.seed() in the test scripts
+* Updated license template
 
 ## Test environments
 * Windows 10, R 3.3.3 (local)
