+++ vignettes/ReinforcementLearning.Rmd
@@ -13,7 +13,7 @@ vignette: >
 set.seed(0)
 ```
 
-This vignette gives a short introduction to the `ReinforcementLearning` package, which allows one to perform model-free reinforcement in *R*. The implementation uses input data in the form of sample sequences consisting of states, actions and rewards. Based on such training examples, the package allows a reinforcement learning agent to learn an optimal policy that defines the best possible action in each state. In the following sections, we present multiple step-by-step examples to illustrate how to take advantage of the capabilities of the `ReinforcementLearning` package. Moreover, we present methods to customize the learning and action selection behavior of the agent. Main features of `ReinforcementLearning` include, but are not limited to,
+This vignette gives a introduction to the `ReinforcementLearning` package, which allows one to perform model-free reinforcement in *R*. The implementation uses input data in the form of sample sequences consisting of states, actions and rewards. Based on such training examples, the package allows a reinforcement learning agent to learn an optimal policy that defines the best possible action in each state. In the following sections, we present multiple step-by-step examples to illustrate how to take advantage of the capabilities of the `ReinforcementLearning` package. Moreover, we present methods to customize the learning and action selection behavior of the agent. Main features of `ReinforcementLearning` include, but are not limited to,
 
 - Learning an optimal policy from a fixed set of a priori known transition samples
 - Predefined learning rules and action selection modes
@@ -65,31 +65,33 @@ library(ReinforcementLearning)
 
 The following sections present the usage and main functionality of the `ReinforcementLearning` package.
 
-## Data format
 
-The `ReinforcementLearning` package uses experience replay to learn an optimal policy based on past experience in the form of sample sequences consisting of states, actions and rewards. Here each training example consists of a state transition tuple *(s,a,r,s_new)*, as described ibelow.
+## Data preparation
 
-* **s** The current environment state.
-* **a** The selected action in the current state.
-* **r** The immediate reward received after transitioning from the current state to the next state.
-* **s_new** The next environment state.
+The \pkg{ReinforcementLearning} package utilizes different mechanisms for reinforcement learning, including Q-learning and experience replay. It thereby learns an optimal policy based on past experience in the form of sample sequences consisting of states, actions and rewards. Consequently, each training example consists of a state-transition tuple $(s_i, a_i, r_{i+1}, s_{i+1})$ as follows:
 
-Note:
+* $s_i$ is the current environment state.
+* $a_i$ denotes the selected action in the current state.
+* $r_{i+1}$ specifies the immediate reward received after transitioning from the current state to the next state.
+* $s_{i+1}$ refers to the next environment state.
 
-* The input data must be a dataframe in which each row represents a state transition tuple *(s,a,r,s_new)*.
+The training examples for reinforcement learning can (1)&nbsp;be collected from an external source and inserted into a tabular data structure, or (2)&nbsp;generated dynamically by querying a function that defines the behavior of the environment. In both cases, the corresponding input must follow the same tuple structure $(s_i, a_i, r_{i+1}, s_{i+1})$. We detail both variants in the following.
 
-## Read-in sample experience
+### Learning from pre-defined observations
 
-The state transition tuples can be collected from an external source and easily read-in into *R*. The sample experience can then be used to train a reinforcement learning agent without requiring further interaction with the environment. The following example shows a representative dataset containing game states of 100,000 randomly sampled Tic-Tac-Toe games. 
[…]
