### 2019-05-24 d54f104
* Updated vignette
