### 2019-05-24 63b2499
* Updated JOSS paper
