+++ JOSS/paper.md
@@ -6,17 +6,14 @@ tags:
   - Batch Learning
   - Experience Replay
   - Q-Learning
-authors:
-  - name: Nicolas Proellochs
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+author:
+  - name: Nicolas Pr&ouml;llochs
+    affiliation: University of Giessen, University of Oxford
+    email: \email{nicolas.proellochs@wi.jlug.de}
   - name: Stefan Feuerriegel
-    affiliation: 2
-affiliations:
- - name: University of Oxford
-   index: 1
- - name: ETH Zurich
-   index: 2
-date: 22 October 2018
+    affiliation: ETH Zurich
+    email: \email{sfeuerriegel@ethz.ch}
+date: 24 May 2019
 bibliography: paper.bib
 output:
   html_document:
@@ -25,9 +22,22 @@ output:
 
 # Summary
 
-Reinforcement learning refers to a group of methods from artificial intelligence where an agent performs learning through trial and error [@Sutton.1998]. It differs from supervised learning, since reinforcement learning requires no explicit labels; instead, the agent interacts continuously with its environment. That is, the agent starts in a specific state and then performs an action, based on which it transitions to a new state and, depending on the outcome, receives a reward. Different strategies (e.g. Q-learning) have been proposed to maximize the overall reward, resulting in a so-called policy, which defines the best possible action in each state. As a main advantage, reinforcement learning is applicable to situations in which the dynamics of the environment are unknown or too complex to evaluate [e.g. @Mnih.2015]. 
+Reinforcement learning refers to a group of methods from artificial intelligence where an agent performs learning through trial and error [@Sutton.1998]. It differs from supervised learning, since reinforcement learning requires no explicit labels; instead, the agent interacts continuously with its environment. That is, the agent starts in a specific state and then performs an action, based on which it transitions to a new state and, depending on the outcome, receives a reward. Different strategies (e.g. Q-learning) have been proposed to maximize the overall reward, resulting in a so-called policy, which defines the best possible action in each state. As a main advantage, reinforcement learning is applicable to situations in which the dynamics of the environment are unknown or too complex to evaluate [e.g. @Mnih.2015]. However, there is currently no package available for performing reinforcement learning in R. As a remedy, we introduce the *ReinforcementLearning* R package, which allows an agent to learn the optimal behavior based on sampling experience consisting of states, actions and rewards [@ReinforcementLearning.2017]. Based on such training examples, the package allows a reinforcement learning agent to learn an optimal policy that defines the best possible action in each state. Main features of `ReinforcementLearning` include, but are not limited to,
 
-However, the available tools in R are not yet living up to the needs of users in such cases. In fact, there is currently no package available that allows one to perform model-free reinforcement learning in R. Hence, users that aim to teach optimal behavior through trial-and-error learning must implement corresponding learning algorithms in a manual way. As a remedy, we introduce the *ReinforcementLearning* package for R, which allows an agent to learn the optimal behavior based on sampling experience consisting of states, actions and rewards [@ReinforcementLearning.2017]. The training examples for reinforcement learning can originate from real-world applications, such as sensor data. In addition, the package is shipped with the built-in capability to sample experience from a function that defines the dynamics of the environment. In both cases, the result of the learning process is a highly interpretable reinforcement learning policy that defines the best possible action in each state. The package provides a remarkably flexible framework, which makes it readily applicable to a wide range of different problems. 
+- Learning an optimal policy from a fixed set of a priori known transition samples
+- Predefined learning rules and action selection modes
+- A highly customizable framework for model-free reinforcement learning tasks
+
+# Statement of Need
+
+<!--Model-based RL / MDPs --> 
[…]
