+++ JOSS/paper.bib
@@ -1,4 +1,4 @@
-﻿
+
 @article{Bellman.1957,
  author = {Bellman, Richard},
  year = {1957},
@@ -26,7 +26,8 @@
  year = {2017},
  title  = {MDPtoolbox},
  note   = {R~package version~4.0-3},
- url = {https://cran.r-project.org/web/packages/MDPtoolbox/index.html}
+ url = {https://cran.r-project.org/web/packages/MDPtoolbox/index.html},
+ doi = {10.1111/ecog.00888}
 }
 
 
@@ -124,7 +125,8 @@
  pages = {529--533},
  volume = {518},
  number = {7540},
- journal = {Nature}
+ journal = {Nature},
+ DOI = {10.1038/nature14236}
 }
 
 
@@ -182,7 +184,8 @@
  title = {Reinforcement Learning: An Introduction},
  address = {Cambridge, MA},
  publisher = {{MIT Press}},
- series = {Adaptive Computation and Machine Learning}
+ series = {Adaptive Computation and Machine Learning},
+ doi = {10.1109/TNN.1998.712192}
 }
 
 
