### 2019-05-24 12f9a5a
* Added NEWS * Improved documentation of package functions
