### 2019-05-24 48a2f29
* Added notes on performance
