+++ JOSS/paper.md
@@ -37,7 +37,7 @@ Reinforcement learning techniques can primarily be categorized in two groups, na
 The second category of reinforcement learning comprises \emph{model-free} approaches that forgo any explicit knowledge regarding the dynamics of the environment. These approaches learn the optimal behavior through trial-and-error by directly interacting with the environment. In this context, the learner has no explicit knowledge of either the reward function or the state transition function [@Sutton.1998]. Instead, the optimal behavior is iteratively inferred from the consequences of actions on the basis of past experience. As a main advantage, this method is applicable to situations in which the dynamics of the environment are unknown or too complex to evaluate. However, the available tools in R are not yet living up to the needs of users in such cases. In fact, there is currently no package available that allows one to perform model-free reinforcement learning in R. Hence, users that aim to teach optimal behavior through trial-and-error learning must implement corresponding learning algorithms in a manual way.
 
 <!-- RL package --> 
-As a remedy, we introduce the \pkg{ReinforcementLearning} package for R, which allows an agent to learn the optimal behavior based on sampling experience consisting of states, actions and rewards  [@ReinforcementLearning.2017]. The training examples for reinforcement learning can originate from real-world applications, such as sensor data. In addition, the package is shipped with the built-in capability to sample experience from a function that defines the dynamics of the environment. In both cases, the result of the learning process is a highly interpretable reinforcement learning policy that defines the best possible action in each state. The package provides a remarkably flexible framework, which makes it readily applicable to a wide range of different problems. Among other functions, it allows one to customize a variety of learning parameters and elaborates on how to mitigate common performance in common solution algorithms (e.g. experience replay). The package vignette demostrates its use by drawing upon common examples from the literature (e.g. finding optimal game strategies).
+As a remedy, we introduce the \pkg{ReinforcementLearning} package for R, which allows an agent to learn the optimal behavior based on sampling experience consisting of states, actions and rewards  [@ReinforcementLearning.2017]. The training examples for reinforcement learning can originate from real-world applications, such as sensor data. In addition, the package is shipped with the built-in capability to sample experience from a function that defines the dynamics of the environment. In both cases, the result of the learning process is a highly interpretable reinforcement learning policy that defines the best possible action in each state. The package provides a remarkably flexible framework, which makes it readily applicable to a wide range of different problems. Among other functions, it allows one to customize a variety of learning parameters and elaborates on how to mitigate common performance in common solution algorithms (e.g. experience replay). The package [vignette](https://github.com/nproellochs/ReinforcementLearning/blob/master/vignettes/ReinforcementLearning.Rmd) demonstrates its use by drawing upon common examples from the literature (e.g. finding optimal game strategies).
 
 # Functionality
 
@@ -64,8 +64,21 @@ environment <- function(state, action) {
 
 After specifying the environment function, one can use *sampleExperience()* to collect random sequences from it. Thereby, the input specifies number of samples ($N$), the environment function, the set of states (i.e. $S$) and the set of actions (i.e. $A$). The return value is then a data frame containing the experienced state transition tuples $(s_i, a_i, r_{i+1}, s_{i+1})$ for $i = 1, \ldots, N$. 
 
-# Notes on Performance
 
-TBC
+# Notes on performance
+
[…]
