### 2019-05-26 #1087 @nproellochs
Thank you again @kbenoit  & @terrytangyuan for reviewing the package. As part of the revision, I have addressed the checklist issues as follows:

>  Version: Does the release version given match the GitHub release (v1.0.2)?

I added a tagged GitHub release matching the CRAN version (v1.0.4).

>  Functionality: Have the functional claims of the software been confirmed?

I have significantly improved the package documentation. In addition, the vignette now contains two working examples (adapted from Sutton (1998)) with code executions and detailed explanations of the data format and model outcomes.

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I added a statement of need section to the paper. In addition, the paper (and the package vignette) now contains a dedicated section addressing the performance of the package. Specifically, I added examples of things the package can do and the things the package cannot do (where users may consider implementations written in "faster" programming languages)

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

As aforementioned, I implemented two working examples in the package vignette. Specifically, I demonstrate the use of the package by drawing upon common examples from the literature (e.g. finding optimal game strategies).

>   Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

See points above. In addition, I added further documentation and additional examples to the manual entries.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I added community guidelines in the README.

> Software paper

I added a statement of need highlighting our contributions to the existing landscape of R packages. In addition, the paper now includes the package authors' affiliations and DOIs.

> Minor issues

I added further documentation and additional examples to the manual entries. In addition, I improved the consistency of function naming and added spaces after `if`.

> R CMD CHECK issues

Fixed.
