### 2019-05-26 #1087 @arfon
Thanks for the update @nproellochs. @terrytangyuan @kbenoit - could you take a look at the updated paper ☝️and let me know if this satisfies your concerns about the usefulness of this software given both of your concerns about performance?

As a reminder, our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) state that:

> - The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

From my perspective, if this is the first reinforcement learning package in R, then that _could_ be a 'substantial contribution' almost by definition, however, if the performance is so poor that its not useful for all but a very small number of applications, then this may be a moot point.
