### 2019-06-10 #1087 @labarba
I see that both reviewers have concerns about the purpose and need of this software…
https://github.com/openjournals/joss-reviews/issues/1087#issuecomment-453893878
https://github.com/openjournals/joss-reviews/issues/1087#issuecomment-461296889

... and @arfon's assessment is:

> if this is the first reinforcement learning package in R, then that could be a 'substantial contribution' almost by definition, however, if the performance is so poor that its not useful for all but a very small number of applications, then this may be a moot point.

The author says they have expanded in the paper the sections on Statement of Need, and Notes on Performance.

Does this new text in the paper satisfy the reviewers?
