### 2019-06-11 #1087 @terrytangyuan
Then see my very first comment https://github.com/openjournals/joss-reviews/issues/1087#issuecomment-443436417, it is **not** the first reinforcement learning package in R.
