### 2019-06-11 #1087 @nproellochs
@terrytangyuan Apart from the abovementioned differences regarding package features, I want to point out that the reinforcelearn R package was released approximately 11 months **after** the ReinforcementLearning package. Therefore, in my opinion, it is fair to claim that the ReinforcementLearning package is the first package that performs model-free RL in R.
