### 2019-06-12 #1087 @arfon
@terrytangyuan @kbenoit - many thanks for all of your input to date and your excellent reviews. While there isn't clear agreement here between us, I'm satisfied with the responses of the author and am making the editorial decision to accept this submission into JOSS.
