### 2019-06-12 #1087 @nproellochs
Thank you @arfon and both reviewers! The archive is now available on Zenodo [10.5281/zenodo.3244170](https://doi.org/10.5281/zenodo.3244170).
