### 2019-06-12 #1087 @arfon
@nproellochs - could you check if any of the DOIs listed by @whedon [above](https://github.com/openjournals/joss-reviews/issues/1087#issuecomment-501326964) are the correct DOIs for the papers in your bibtex file? If they are correct, please add them to you bibtex file.
