### 2019-06-12 bb0b982
* Cleaned-up bibtex file
