+++ JOSS/paper.bib
@@ -1,26 +1,4 @@
 
-@article{Bellman.1957,
- author = {Bellman, Richard},
- year = {1957},
- title = {A Markovian Decision Process},
- pages = {679--684},
- volume = {6},
- number = {5},
- journal = {Journal of Mathematics and Mechanics}
-}
-
-
-@article{Beom.1995,
- author = {Beom, Hee Rak and Cho, Hyung Suck},
- year = {1995},
- title = {A Sensor-Based Navigation for a Mobile Robot using Fuzzy Logic and Reinforcement Learning},
- pages = {464--477},
- volume = {25},
- number = {3},
- journal = {IEEE Transactions on Systems, Man, and Cybernetics},
-}
-
-
 @Manual{MDPtoolbox.2017,
  author = {Chades, Iadine and Chapron, Guillaume and Cros, Marie-Josee and Garcia, Frederick and Sabbadin, Regis},
  year = {2017},
@@ -28,93 +6,6 @@
  note   = {R~package version~4.0-3},
  url = {https://cran.r-project.org/web/packages/MDPtoolbox/index.html},
  doi = {10.1111/ecog.00888}
-}
-
-
-@article{Barto.1996,
- author = {Crites, R. H. and Barto, Andrew G.},
- year = {1996},
- title = {Improving Elevator Performance using Reinforcement Learning},
- pages = {1017--1023},
- volume = {8},
- journal = {Advances in Neural Information Processing Systems}
-}
-
-
-@article{Erev.1998,
- author = {Erev, Ido and Roth, Alvin E.},
- year = {1998},
- title = {Predicting How People Play Games: Reinforcement Learning in Experimental Games with Unique, Mixed Strategy Equilibria},
- pages = {848--881},
- journal = {American Economic Review}
-}
-
-
-@article{Erev.1998b,
- author = {Erev, Ido and Rapoport, Amnon},
- year = {1998},
- title = {Coordination, ``Magic,'' and Reinforcement Learning in a Market Entry Game},
- pages = {146--175},
- volume = {23},
- number = {2},
- journal = {Games and Economic Behavior}
-}
-
-
-@article{Hu.2003,
- author = {Hu, Junling and Wellman, Michael P.},
- year = {2003},
- title = {Nash Q-Learning for General-Sum Stochastic Games},
- pages = {1039--1069},
- volume = {4},
- journal = {Journal of Machine Learning Research}
-}
-
-
-@article{Kaelbling.1996,
- author = {Kaelbling, Leslie Pack and Littman, Michael L. and Moore, Andrew W.},
- year = {1996},
- title = {Reinforcement Learning: A Survey},
- pages = {237--285},
- volume = {4},
- journal = {Journal of Artificial Intelligence Research}
-}
-
-
-@incollection{Lange.2012,
- author = {Lange, Sascha and Gabel, Thomas and Riedmiller, Martin},
- title = {Batch Reinforcement Learning},
- pages = {45--73},
- volume = {12},
- publisher = {Springer},
- series = {Adaptation, Learning, and Optimization},
- editor = {Wiering, Marco and {van Otterlo}, Martijn},
- booktitle = {Reinforcement Learning},
- year = {2012},
- address = {Heidelberg and New York}
-}
-
-
-@article{Lin.1992,
- author = {Lin, Long-Ji},
- year = {1992},
- title = {Self-Improving Reactive Agents Based on Reinforcement Learning, Planning and Teaching},
- pages = {293--321},
- volume = {8},
- number = {3},
- journal = {Machine Learning}
-}
-
-
-@incollection{Mataric.1997,
- author = {Matari{\'c}, Maja J.},
[…]
