### 2019-06-12 #1087 @arfon
@nproellochs - this should work for your name: https://github.com/nproellochs/ReinforcementLearning/pull/11
