### 2019-06-12 #1087 @nproellochs
@arfon The bib file is fixed now.
